+++ Add-Debugger.ps1
@@ -1,5 +1,5 @@
 <#PSScriptInfo
-.VERSION 1.1.0
+.VERSION 1.2.0
 .AUTHOR Roman Kuzmin
 .COPYRIGHT (c) Roman Kuzmin
 .TAGS Debug
@@ -10,17 +10,16 @@
 
 <#
 .Synopsis
-	Adds a script debugger to PowerShell.
+	Debugger for PowerShell hosts with no own debuggers.
 
 .Description
-	This script is for PowerShell hosts with no own debuggers, e.g. Visual
-	Studio NuGet console ("Package Manager Host"), the default runspace host
-	("Default Host"). But it may be used instead of the default debuggers as
-	well ("ConsoleHost").
+	The script adds or replaces existing debugger in any PowerShell runspace.
+	It is useful for hosts with no own debuggers, e.g. Visual Studio NuGet
+	console ("Package Manager Host"), bare runspace host ("Default Host").
+	And it may be used instead of existing debuggers ("ConsoleHost").
 
-	The script is called once at any moment when debugging is needed. In order
-	to restore the original debugger, invoke Restore-Debugger which is defined
-	by Add-Debugger.
+	The script is called at any moment when debugging is needed. To restore
+	the original debuggers, invoke Restore-Debugger defined by Add-Debugger.
 
 	For output of debug commands the script uses Out-Host or a file with a
 	separate console started for watching its tail.
@@ -30,11 +29,14 @@
 
 	PowerShell commands are invoked in a child of the current scope. In order
 	to change current scope variables you would use `Set-Variable -Scope 1`.
-	But the script recognises `$name = ...` and assigns the "proper" $name.
+	But the script recognises `$var = ...` and assigns in the proper scope.
 
 .Parameter Path
 		Specifies the output file used instead of Out-Host.
 		A separate console is opened for watching its tail.
+
+.Parameter Context
+		Specifies the number of context source lines to show.
 
 .Parameter XPos
 		Horizontal position of the input box.
@@ -61,11 +63,11 @@
 		# add debugger with file output
 		Add-Debugger $env:TEMP\debug.log
 
-		# trick: stop debugger on terminating errors
+		# trick: break on terminating errors
 		$null = Set-PSBreakpoint -Variable StackTrace -Mode Write
 
-		# from now on the debugger dialog is shown on failures
-		# and a separate PowerShell output console is started
+		# from now on the debugger dialog is shown on errors
+		# and a separate PowerShell output console is opened
 		...
 	})
 	$ps.Invoke() # or BeginInvoke()
@@ -77,6 +79,7 @@
 param(
 	[Parameter()]
 	[string]$Path,
+	[int]$Context,
 	[int]$XPos = -1,
 	[int]$YPos = -1,
 	[switch]$ReadHost
@@ -119,14 +122,14 @@ function global:Restore-Debugger {
 ### Set debugger data.
 $null = New-Variable -Name _Debugger -Scope Global -Description Add-Debugger.ps1 -Option ReadOnly -Value @{
 	Path = $null
-	XPos = $XPos
-	YPos = $YPos
+	Context = $Context
+	XPos = [ref]$XPos
+	YPos = [ref]$YPos
 	Args = $null
 	Watch = $false
 	History = [System.Collections.ArrayList]@()
 	Handlers = Remove-Debugger
-	DefaultContext = 0
-	Context = [ref]0
+	RefContext = [ref]0
 	Action = '?'
 	Match = $null
 }
@@ -137,7 +140,7 @@ if ($Path) {
 	$_Debugger.Watch = $true
 	function global:Write-Debugger {
 		param($Data)
-		[System.IO.File]::AppendAllText($_Debugger.Path, ($Data | Out-String), [System.Text.Encoding]::Unicode)
+		[System.IO.File]::AppendAllText($_Debugger.Path, ($Data | Out-String))
 		if ($_Debugger.Watch) {
 			$_Debugger.Watch = $false
 			Watch-Debugger $_Debugger.Path
@@ -159,10 +162,73 @@ if ($ReadHost) {
 	}
 }
 else {
-	Add-Type -AssemblyName Microsoft.VisualBasic
 	function global:Read-Debugger {
 		param($Prompt, $Default)
-		[Microsoft.VisualBasic.Interaction]::InputBox($Prompt, 'Add-Debugger', $Default, $_Debugger.XPos, $_Debugger.YPos)
+		Read-InputBox $Prompt Add-Debugger $Default Step Continue $_Debugger.XPos $_Debugger.YPos
+	}
+}
+
+# Gets an input string from a dialog.
+function global:Read-InputBox {
+	param($Prompt, $Title, $Default, $Text1, $Text2, $XPos, $YPos)
+
+	Add-Type -AssemblyName System.Windows.Forms, System.Drawing
+
+	$form = New-Object System.Windows.Forms.Form
+	$form.Text = $Title
+	$form.TopMost = $true
+	$form.Size = New-Object System.Drawing.Size(400, 132)
+	$form.FormBorderStyle = 'FixedDialog'
+	if ($XPos -and $YPos -and $XPos.Value -ge 0 -and $YPos.Value -ge 0) {
+		$form.StartPosition = 'Manual'
+		$form.Location = New-Object System.Drawing.Point($XPos.Value, $YPos.Value)
+	}
+	else {
+		$form.StartPosition = 'CenterScreen'
+	}
+
+	$label = New-Object System.Windows.Forms.Label
+	$label.Location = New-Object System.Drawing.Point(10, 10)
+	$label.Size = New-Object System.Drawing.Size(380, 20)
+	$label.Text = $Prompt
+	$form.Controls.Add($label)
+
+	$text = New-Object System.Windows.Forms.TextBox
+	$text.Text = $Default
+	$text.Location = New-Object System.Drawing.Point(10, 30)
+	$text.Size = New-Object System.Drawing.Size(365, 20)
+	$form.Controls.Add($text)
+
+	$ok = New-Object System.Windows.Forms.Button
+	$ok.Location = New-Object System.Drawing.Point(225, 60)
+	$ok.Size = New-Object System.Drawing.Size(75, 23)
+	$ok.Text = $Text1
+	$ok.DialogResult = 'OK'
+	$form.AcceptButton = $ok
+	$form.Controls.Add($ok)
+
+	$cancel = New-Object System.Windows.Forms.Button
+	$cancel.Location = New-Object System.Drawing.Point(300, 60)
+	$cancel.Size = New-Object System.Drawing.Size(75, 23)
+	$cancel.Text = $Text2
+	$cancel.DialogResult = 'Cancel'
+	$form.CancelButton = $cancel
+	$form.Controls.Add($cancel)
+
+	$form.add_Load({
+		$text.Select()
+		$form.Activate()
+	})
+
+	$result = $form.ShowDialog()
+
+	if ($XPos -and $YPos) {
+		$XPos.Value = $form.Location.X
+		$YPos.Value = $form.Location.Y
+	}
+
+	if ($result -eq 'OK') {
+	    $text.Text
 	}
 }
 
@@ -170,9 +236,8 @@ else {
 function global:Watch-Debugger {
 	param($Path)
 	$Path = $Path.Replace("'", "''")
-	$me = Get-Process -Id $PID
-	$app = if ($me.Name -eq 'pwsh') {$me.Path} else {'powershell.exe'}
-	Start-Process $app "-NoProfile -Command Get-Content -LiteralPath '$Path' -Wait -ErrorAction 0"
+	$app = if ($PSVersionTable.PSEdition -eq 'Core' -and (Get-Command pwsh -ErrorAction 0)) {'pwsh'} else {'powershell'}
+	Start-Process $app "-NoProfile -Command `$Host.UI.RawUI.WindowTitle = 'Debug output'; Get-Content -LiteralPath '$Path' -Encoding UTF8 -Wait"
 }
 
 # Writes the current invocation info.
@@ -244,49 +309,64 @@ function global:Write-DebuggerFile {
 	}}
 
 	# write debug location
-	Write-DebuggerInfo $_.InvocationInfo $_Debugger.DefaultContext
+	Write-DebuggerInfo $_.InvocationInfo $_Debugger.Context
 	Write-Debugger ''
 
-	# REPL
+	# restore parent $_
 	$_Debugger.Args = $_
+	if ($_ = Get-Variable [_] -Scope 1 -ErrorAction 0) {
+		$_ = $_.Value
+	}
+
+	# REPL
 	for() {
 		### prompt
-		$_Debugger.Action = Read-Debugger "Enter PowerShell and debug commands.`nUse h or ? for help" $_Debugger.Action
-		$_Debugger.Action = if ($null -eq $_Debugger.Action) {''} else {$_Debugger.Action.Trim()}
+		$_Debugger.LastAction = $_Debugger.Action
+		$_Debugger.Action = Read-Debugger 'Step (h or ? for help)' $_Debugger.Action
+		if ($_Debugger.Action) {
+			$_Debugger.Action = $_Debugger.Action.Trim()
+		}
 		Write-Debugger "DBG> $($_Debugger.Action)"
 
+		### repeat
+		if ($_Debugger.Action -eq '' -and ('s', 'StepInto', 'v', 'StepOver') -contains $_Debugger.LastAction) {
+			$_Debugger.Action = $_Debugger.LastAction
+			$_Debugger.Args.ResumeAction = if (('s', 'StepInto') -contains $_Debugger.Action) {'StepInto'} else {'StepOver'}
+			return
+		}
+
 		### Continue
-		if (!$_Debugger.Action -or $_Debugger.Action -eq 'c' -or $_Debugger.Action -eq 'Continue') {
+		if (($null, 'c', 'Continue') -contains $_Debugger.Action) {
 			$_Debugger.Args.ResumeAction = 'Continue'
 			return
 		}
 
 		### StepInto
-		if ($_Debugger.Action -eq 's' -or $_Debugger.Action -eq 'StepInto') {
+		if (('s', 'StepInto') -contains $_Debugger.Action) {
 			$_Debugger.Args.ResumeAction = 'StepInto'
 			return
 		}
 
 		### StepOver
-		if ($_Debugger.Action -eq 'v' -or $_Debugger.Action -eq 'StepOver') {
+		if (('v', 'StepOver') -contains $_Debugger.Action) {
 			$_Debugger.Args.ResumeAction = 'StepOver'
 			return
 		}
 
 		### StepOut
-		if ($_Debugger.Action -eq 'o' -or $_Debugger.Action -eq 'StepOut') {
+		if (('o', 'StepOut') -contains $_Debugger.Action) {
 			$_Debugger.Args.ResumeAction = 'StepOut'
 			return
 		}
 
 		### Quit
-		if ($_Debugger.Action -eq 'q' -or $_Debugger.Action -eq 'Quit') {
+		if (('q', 'Quit') -contains $_Debugger.Action) {
 			$_Debugger.Args.ResumeAction = 'Stop'
 			return
 		}
 
 		### Detach
-		if ($_Debugger.Action -eq 'd' -or $_Debugger.Action -eq 'Detach') {
+		if (('d', 'Detach') -contains $_Debugger.Action) {
 			if ($_Debugger.Handlers) {
 				Write-Debugger 'd, Detach - not supported in this environment.'
 				continue
@@ -298,32 +378,33 @@ function global:Write-DebuggerFile {
 		}
 
 		### history
-		if ($_Debugger.Action -eq 'r') {
+		if ('r' -eq $_Debugger.Action) {
 			Write-Debugger $_Debugger.History
 			continue
 		}
 
 		### stack
-		if ($_Debugger.Action -ceq 'k') {
+		if ('k' -ceq $_Debugger.Action) {
 			Write-Debugger (Get-PSCallStack | Format-Table Command, Location, Arguments -AutoSize)
 			continue
 		}
-		if ($_Debugger.Action -ceq 'K') {
+		if ('K' -ceq $_Debugger.Action) {
 			Write-Debugger (Get-PSCallStack | Format-List)
 			continue
 		}
 
 		### <number>
-		if ([int]::TryParse($_Debugger.Action, $_Debugger.Context)) {
-			Write-DebuggerInfo $_Debugger.Args.InvocationInfo $_Debugger.Context.Value
+		if ([int]::TryParse($_Debugger.Action, $_Debugger.RefContext)) {
+			Write-DebuggerInfo $_Debugger.Args.InvocationInfo $_Debugger.RefContext.Value
 			if ($_Debugger.Action[0] -eq "+") {
-				$_Debugger.DefaultContext = $_Debugger.Context.Value
+				$_Debugger.Context = $_Debugger.RefContext.Value
 			}
+			$_Debugger.Action = $_Debugger.LastAction
 			continue
 		}
 
 		### watch
-		if ($_Debugger.Action -eq 'w') {
+		if ('w' -eq $_Debugger.Action) {
 			if ($_Debugger.Path) {
 				Watch-Debugger $_Debugger.Path
 			}
@@ -334,18 +415,16 @@ function global:Write-DebuggerFile {
 		}
 
 		### help
-		if ($_Debugger.Action -eq '?' -or $_Debugger.Action -eq 'h') {
+		if (('?', 'h') -contains $_Debugger.Action) {
 			Write-Debugger (@(
 				''
 				'  s, StepInto  Step to the next statement into functions, scripts, etc.'
 				'  v, StepOver  Step to the next statement over functions, scripts, etc.'
 				'  o, StepOut   Step out of the current function, script, etc.'
-				'  c, Continue  Continue operation (also on empty input).'
-
+				'  c, Continue  Continue operation.'
 				if (!$_Debugger.Handlers) {
-					'  d, Detach    Continue operation and detach the debugger.'
+				'  d, Detach    Continue operation and detach the debugger.'
 				}
-
 				'  q, Quit      Stop operation and exit the debugger.'
 				'  ?, h         Write this help message.'
 				'  k            Write call stack (Get-PSCallStack).'
@@ -355,8 +434,11 @@ function global:Write-DebuggerFile {
 				'  +<n>         Set location context preference to <n> lines.'
 				'  k <s> <n>    Write source at stack <s> in context of <n> lines.'
 				''
+				if ($_Debugger.Path) {
 				'  w            Restart watching the debugger output file.'
+				}
 				'  r            Write last PowerShell commands invoked on debugging.'
+				'  <empty>      Repeat the last command if it was StepInto, StepOver.'
 				'  <command>    Invoke any PowerShell <command> and write its output.'
 				''
 			) -join [System.Environment]::NewLine)